fixed issues

+++ install.sql
@@ -102,12 +102,12 @@ cumulativeLikes INT(10) NOT NULL DEFAULT 0
 DROP TABLE IF EXISTS cms1_counter;
 CREATE TABLE cms1_counter(
 
-time INT (10) NOT NULL,
+time INT (20) NOT NULL,
 userID INT(10),
 browser VARCHAR(255),
 browserVersion VARCHAR(255),
 ipAddress VARCHAR(255),
-spider TINYINT(1)
+spider INT(20)
 );
 
 --news to category
